se agregaron los controles necesarios en las acciones que se pueden realizar en el views de proyectos, y una partecita de fases!

+++ SGP_Q03/proyectos/views.py
@@ -11,6 +11,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.forms import ModelForm
 from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Submit, Button
+from usuarios.models import Usuario
 
 from .models import Proyecto
 """ login/views
@@ -51,29 +52,63 @@ def proyecto_list(request, template_name='proyectos/proyecto_list.html'):
     data['object_list'] = proyectos
     return render(request, template_name, data)
 
+""" Función: proyecto_create
+    No recibe parámetros
+    Descripción: Solo el usuario administrador del sistema puede crear un proyecto"""
 @login_required
 def proyecto_create(request, template_name='proyectos/proyecto_form.html'):
     form = ProyectoForm(request.POST or None)
-    if form.is_valid():
-        #f = Proyecto(nombre=request.POST['nombre'], estado="Pendiente",) #estado pendiente por defecto
-        #f.save()
-        form.save()
-        return redirect('lista_proyecto')
-    return render(request, template_name, {'form': form})
+    usuario_administrador=request.user.is_superuser
+    if usuario_administrador is True:
+        if form.is_valid():
+            #f = Proyecto(nombre=request.POST['nombre'], estado="Pendiente",) #estado pendiente por defecto
+            #f.save()
+            form.save()
+            return redirect('lista_proyecto')
+        return render(request, template_name, {'form': form})
+    else:
+        data={}
+        data['mensaje'] = "Lo sentimos, Usted no es lider de este proyecto"
+        return render(request, 'proyectos/proyecto_list.html', data)
 
+""" Función: proyecto_update
+    Parámetros que recibe
+    @param: pk
+    @value: es el primary key del proyecto
+    Descripción: Solo los usuarios Lider o el administrador del proyecto pueden editar las configuraciones iniciales del proyecto, siempre y cuando
+    el estado del proyecto sea Pendiente"""
 @login_required
 def proyecto_update(request, pk, template_name='proyectos/proyecto_form.html'):
     proyecto = get_object_or_404(Proyecto, pk=pk)
     form = ProyectoForm(request.POST or None, instance=proyecto)
-    if form.is_valid():
-        form.save()
-        return redirect('lista_proyecto')
-    return render(request, template_name, {'form': form})
+    request_user = request.user.id
+    usuario_administrador=request.user.is_superuser
 
+    usuario_lider = Usuario.objects.get(pk=proyecto.lider.pk) #este es el usuario lider
+
+    if proyecto.estado == "Pendiente" and request_user == usuario_lider.pk+1 or proyecto.estado == "Pendiente" and usuario_administrador is True: #si el estado es pendiente y si el lider o el admin son los que acceden
+        if form.is_valid():
+            form.save()
+            return redirect('lista_proyecto')
+        return render(request, template_name, {'form': form})
+    else:
+
+        return render(request, 'proyectos/proyecto_sin_permisos.html', {})
+
+""" Función: proyecto_delete
+    Parametros que recibe
+    @param: pk
+    @value: es el primary key del proyecto
+    Descripción: Solo el usuario administrador del proyecto puede eliminar el proyecto siempre y cuando
+    el estado sea Pendiente"""
 @login_required
 def proyecto_delete(request, pk, template_name='proyectos/proyecto_confirm_delete.html'):
     server = get_object_or_404(Proyecto, pk=pk)
-    if request.method=='POST':
-        server.delete()
-        return redirect('lista_proyecto')
-    return render(request, template_name, {'object': server})
+    usuario_administrador=request.user.is_superuser
+    if server.estado == "Pendiente" and usuario_administrador is True:
+        if request.method=='POST':
+            server.delete()
+            return redirect('lista_proyecto')
+        return render(request, template_name, {'object': server})
+    else:
+        return render(request, 'proyectos/proyecto_sin_permisos.html', {})
